scraper/requirements.txt
# Python dependencies for Facebook Ads scraper
selenium>=4.15.0
webdriver-manager>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
